[WIP] Finish hooking up rest of demo for WinUI app with separate ChatService
Still some clean-up/tidying can be done especially with a bit of passing of stuff around that maybe DependencyInjection could help with.
Also, not really dealing with overall status, but main visible bits are good.
TODO: Need to clean-up the visual on the overlay for the processing (grab brushes from WPF stuff I had done earlier for that).